Split `SourceErrorFormatter` type from `PlainTextFormatter`.
