+++ src/formatter.rs
@@ -1,5 +1,10 @@
 //! Formatters for source errors.
 
-pub use self::plain_text_formatter::PlainTextFormatter;
+pub use self::{
+    plain_text_formatter::PlainTextFormatter, source_error_formatter::SourceErrorFormatter,
+    styler::Styler,
+};
 
 mod plain_text_formatter;
+mod source_error_formatter;
+mod styler;
